Changes to deleted posts timestamp storage, initial 2038 compliance.

+++ src/main/java/net/easymodo/asagi/DumperJSON.java
@@ -70,7 +70,7 @@ public class DumperJSON extends AbstractDumper {
                 for(Topic oldTopic : topics.values()) {
                     oldTopic.lock.readLock().lock();
                     int oldTopicNum = oldTopic.getNum();
-                    int oldTopicLastMod = oldTopic.getLastModTimestamp();
+                    long oldTopicLastMod = oldTopic.getLastModTimestamp();
                     oldTopic.lock.readLock().unlock();
 
                     Topic newTopic = threadMap.remove(oldTopicNum);
